Removed blanket except that is supposed to catch SocketTimeoutErrors; unrelated cleanup.
This was implemented as a stopgap solution for SeattleTestbed/portability#28.
A less error prone variant would have been to check `str(exceptionobject)`
for the name of the expected exception; the blanket except has been
dening me around in circles during debugging...

I'm disabling it for now. I'll implement the suggested name-in-str
fix until the root cause with `dylink` is fixed.