Work in progress: Get the inclusion of Affix right. Also, unrelated cleanup.
Lots of the overriding etc. stuff has been on this branch before,
but I believe it was never tested. E.g., the nodemanager's listening
socket was wrapped using `sockettimeout`, but the wrong wrapper.

Other than that, added a few debug statements to see which network
calls are actually used, and added a few spaces between operators
here and there.