Remove debug messages; check that the accepter thread is running (and restart it if not).
Previously, this restarting functionality was implemented in the
check for IP address changes. Now that NamingAndResolverAffix
does the check, we need a different way to find we should restart
AccepterThread. Also, I don't see that previously, an AccepterThread
that became invalid (e.g. TCPServerSocket not usable anymore due to
IP address change) would ever be stopped.